TODO-721: moving stuff from CLI to Util
