+++ content/OTRadioLink/utility/OTV0P2BASE_CLI.cpp
@@ -20,21 +20,6 @@ Author(s) / Copyright (s): Deniz Erbilgin 2016
 
 namespace OTV0P2BASE {
 
-/**
- * @brief   Convert 2 hex characters into a a binary value
- * @param   pointer to a token containing characters between 0-9, a-f or A-F
- * @retval  byte containing converted value
- * @todo    write good comments
- *          add checks
- */
-uint8_t parseHex(const uint8_t *tok)
-{
-  uint8_t hiNibble = *tok++;
-  uint8_t lowNibble = *tok;
 
-  hiNibble = parseHexVal(hiNibble) << 4;
-  hiNibble |= parseHexVal(lowNibble);
-  return hiNibble;
-}
 
 }
