+++ content/OTRadioLink/utility/OTV0P2BASE_Util.cpp
@@ -16,20 +16,29 @@ under the Licence.
 Author(s) / Copyright (s): Deniz Erbilgin 2016
 */
 
-// CLI support routines
-
-#ifndef CONTENT_OTRADIOLINK_UTILITY_OTV0P2BASE_CLI_H_
-#define CONTENT_OTRADIOLINK_UTILITY_OTV0P2BASE_CLI_H_
-
-#include <stdint.h>
+#include "OTV0P2BASE_Util.h"
 
 namespace OTV0P2BASE {
 
 
+/**
+ * @brief   Convert 2 hex characters (eg "0a") into a binary value
+ * @param   pointer to a token containing characters between 0-9, a-f or A-F
+ * @retval  byte containing converted value
+ * @todo    write good comments
+ *          add checks
+ */
+uint8_t parseHex(const uint8_t *tok)
+{
+  if((NULL == tok) || ('\0' == tok[0]) || ('\0' == tok[1])) { return(0); } // ERROR
 
+  uint8_t hiNibble = *tok++;
+  uint8_t lowNibble = *tok;
 
+  hiNibble = parseHexVal(hiNibble) << 4;
+  hiNibble |= parseHexVal(lowNibble);
+  return hiNibble;
 }
 
 
-
-#endif /* CONTENT_OTRADIOLINK_UTILITY_OTV0P2BASE_CLI_H_ */
+}
